version: Support single override and add git hash.
This reworks the way versions are handled internally to match other
Decred software.

In particular, it reverses the semantics such that the individual semver
components (major, minor, patch, prerelease, and buildmetadata) are
parsed from a full string and exported at init time.

Also, since the version is now parsed and verified to be accurate, it
updates the pre-release parsing to properly support dots as required by
the spec.

It adds the git commit hash to the version string as buildmetadata when
no buildmetadata is otherwise specified.

This provides a few main benefits:

- Allows a single linker override to fully specify the version string
  instead of having separate ones that can only override the prerelease
  and build metadata portions
- Provides run-time checks to ensure the full version string is valid
  per the semver spec regardless of whether it was specified directly in
  the source or provided via the linker
- The exact commit used to build non-release versions will be in the
  version string by default

Finally, while here, add some comments regarding the release process to
help maintainers.